requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-mail==0.10.0
apscheduler==3.10.4
psycopg2-binary==2.9.9
pytz==2024.1
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
